==> notebooks/01 preprocessing.ipynb <==
This notebook performs initial preprocessing of three sets of data 1. Annotated data 1 (ca.1000 tweets) 2.Annotated data 2 (ca.500 tweets) 3. Non annotated data (ca.12000 tweets). Schizophrenia data is saved down in the data/dataOut/schiz/ path. The annotated data has the class label 1 removed to convert this into a binary task.

import pandas as pd
import numpy as np

# Twitter Scizhophrenia stigma

'''
import csv file and return dataframe, can
pass column if want to return series
'''
def getFile(fileName, column=None):
    socialDf = pd.read_csv(fileName, encoding='utf-8')
    return socialDf if column==None else socialDf[column]


'''
clean the schizophrenia data. Remove class labels 1 to
return convert into binary classification task. Turn 
label 2 into 1
'''
def cleanScizAnnTwitter(annTwitter):

    print('initially', annTwitter['Classification'].value_counts())
    annTwitter = annTwitter.dropna()
    annTwitter = annTwitter[annTwitter['Classification']!='o']
    annTwitter = annTwitter.astype({"Classification": int})
    annTwitter = annTwitter[annTwitter['Classification']!=1]
    
    annTwitter = annTwitter.replace(2, 1)
    
    print('initially', annTwitter['Classification'].value_counts())
    
    
    return annTwitter

### Generate files for baseline model

fileName = 'data/baseline/dataIn/annFinalSchiz_1.csv'
socialDf = getFile(fileName)
print len(socialDf)
tweets = cleanScizAnnTwitter(socialDf)
print len(tweets), 'grg'
tweets.to_csv('data/baseline/dataOut/annSchiz1.csv', encoding='utf-8')

socialDf['Classification'].value_counts()
tweets['Classification'].value_counts()

fileName = 'data/baseline/dataIn/annFinalSchiz_2.csv'
socialDf = getFile(fileName)
tweets = cleanScizAnnTwitter(socialDf)
tweets.to_csv('data/baseline/dataOut/annSchiz2.csv', encoding='utf-8')

### Generate files for Schizophrenia

fileName = 'data/dataIn/schiz/nonAnnFinalSchiz.csv'
socialDf = getFile(fileName)
socialDf.to_csv('data/dataOut/schiz/nonAnnFinalSchiz.csv', encoding='utf-8')

fileName = 'data/dataIn/schiz/annFinalSchiz_1.csv'
socialDf = getFile(fileName)
tweets = cleanScizAnnTwitter(socialDf)
tweets.to_csv('data/dataOut/schiz/annFinalSchiz_1.csv', encoding='utf-8')
tweets['Classification'].value_counts()

fileName = 'data/dataIn/schiz/annFinalSchiz_2.csv'
socialDf = getFile(fileName)
tweets = cleanScizAnnTwitter(socialDf)
tweets.to_csv('data/dataOut/schiz/annFinalSchiz_2.csv', encoding='utf-8')

### generate files for Stigma Data - Not used in final project

fileName = 'data/dataIn/stigma/nonAnnFinalStig.csv'
socialDf = getFile(fileName)
socialDf.to_csv('data/dataOut/stigma/nonAnnFinalStig.csv', encoding='utf-8')

fileName = 'data/dataIn/stigma/annFinalStig.csv'
socialDf = getFile(fileName)
stigTwitter = socialDf.dropna()
stigTwitter = stigTwitter[socialDf['Classification']!='Neutral']
stigTwitter = stigTwitter[socialDf['Classification']!='Unclear']
stigTwitter = stigTwitter.replace('Anti-stigma', 0)
stigTwitter = stigTwitter.replace('Stigma', 1)

stigTwitter.to_csv('data/dataOut/stigma/annFinalStig.csv', encoding='utf-8')
