==> notebooks/06 bert.ipynb <==
This contains code for pytorch bert implementation. The code was adapted from 
https://mccormickml.com/2019/07/22/BERT-fine-tuning/

from pathlib import Path
  

import numpy as np
import pandas as pd
import torch
from pytorch_pretrained_bert import BertTokenizer
import logging
import tensorflow as tf
from torch.utils.data import TensorDataset, DataLoader, RandomSampler, SequentialSampler
from keras.preprocessing.sequence import pad_sequences
from sklearn.model_selection import train_test_split
from pytorch_pretrained_bert import BertTokenizer, BertConfig
from pytorch_pretrained_bert import BertAdam, BertForSequenceClassification
from tqdm import tqdm, trange
import pandas as pd
import io

device = torch.device("cpu")
tokenizer = BertTokenizer.from_pretrained('bert-large-cased')

fileNameSchiz1 = 'data/dataOut/schiz/annFinalSchiz_1.csv'
fileNameSchiz2 = 'data/dataOut/schiz/annFinalSchiz_2.csv'
fileNameStig = 'data/dataOut/stigma/annFinalStig.csv'

socialDf = pd.read_csv(fileNameSchiz2, encoding='utf-8')
textSchiz2 = socialDf['Tweet']
labelsSchiz2 = socialDf['Classification']

socialDf = pd.read_csv(fileNameSchiz1, encoding='utf-8')
tweetsSchiz1 = socialDf.Tweet.values
labelsSchiz1 = socialDf.Classification.values


Firs we prepare the text in the Bert format with the [CLS] separator and then convert into tensor form. Once we have the data in the right form, including the attention masks we train the model computing a foward pass then using the BertAdam optimizer to minimize our loss. Once this is done we put the model into evaluation mode and evaluate. Can take a little while to run

def getBertFormat(tweets):
    tweets = ['[CLS] ' + s + ' [SEP]' for s in tweets]
    tokens = [tokenizer.tokenize(t) for t in tweets]
    return tokens

def getTorchTensors(**kwargs):
    return {k:torch.tensor(v) for k,v in kwargs.items()}


def getParameters(model):
    
    param_optimizer = list(model.named_parameters())
    
    no_decay = ['bias', 'gamma', 'beta']
    optimizerParameters = [
    {'params': [p for n, p in param_optimizer if not any(nd in n for nd in no_decay)],
     'weight_decay_rate': 0.01},
    {'params': [p for n, p in param_optimizer if any(nd in n for nd in no_decay)],
     'weight_decay_rate': 0.0}]
    
    return optimizerParameters
        
        
def trainModel(model, trainDataLoader, optimizer, epochs=4):
        
    for _ in trange(epochs, desc="Epoch"):
        model.train()
        tr_loss = 0
        nb_tr_examples, nb_tr_steps = 0, 0
  
        for step, batch in enumerate(trainDataLoader):

            batch = tuple(t.to(device) for t in batch)

            b_input_ids, b_input_mask, b_labels = batch

            optimizer.zero_grad()

            loss = model(b_input_ids, token_type_ids=None, attention_mask=b_input_mask, labels=b_labels)   
            loss.backward()
            optimizer.step()

            tr_loss += loss.item()
            nb_tr_examples += b_input_ids.size(0)
            nb_tr_steps += 1

            print("Train loss: {}".format(tr_loss/nb_tr_steps))
        
        
def evalModel(model, optimizer, valDataLoader):
        
    model.eval()
    eval_loss, eval_accuracy = 0, 0
    nb_eval_steps, nb_eval_examples = 0, 0

    for batch in valDataLoader:

        batch = tuple(t.to(device) for t in batch)
        b_input_ids, b_input_mask, b_labels = batch
        
    with torch.no_grad():
        logits = model(b_input_ids, token_type_ids=None, attention_mask=b_input_mask)
    
    # Move logits and labels to CPU
    logits = logits.detach().cpu().numpy()
    label_ids = b_labels.to('cpu').numpy()

    tmp_eval_accuracy = flat_accuracy(logits, label_ids)
    
    eval_accuracy += tmp_eval_accuracy
    nb_eval_steps += 1

    print("Validation Accuracy: {}".format(eval_accuracy/nb_eval_steps))
   
           

def getBert(text, device, batch=16):
    
    tokens = getBertFormat(text)
    #maxLength = max([len(t) for t in tokens]) + 2
    maxLength = 200
    ids = [tokenizer.convert_tokens_to_ids(t) for t in tokens]
    inputIds = pad_sequences(ids, maxlen=maxLength, dtype='long', truncating='post', padding='post')
    
    attentionMasks = [list(map(lambda x: float(x>0), s)) for s in inputIds]
    
    trainInputs, valInputs, trainLabels, valLabels = train_test_split(inputIds, labelsSchiz1, random_state=42, test_size=0.2)
    trainMasks, valMasks, _, _ = train_test_split(attentionMasks, inputIds, random_state=42, test_size=0.2)
    
    inputs = getTorchTensors(trainInputs=trainInputs, valInputs=valInputs, trainLabels=trainLabels, valLabels=valLabels)
    masks = getTorchTensors(trainMasks=trainMasks, valMasks=valMasks)
    
    train_data = TensorDataset(inputs['trainInputs'], masks['trainMasks'], inputs['trainLabels'])
    train_sampler = RandomSampler(train_data)
    train_dataloader = DataLoader(train_data, sampler=train_sampler, batch_size=batch)

    valData = TensorDataset(inputs['valInputs'], masks['valMasks'], inputs['valLabels'])
    valSampler = SequentialSampler(valData)
    valDataLoader = DataLoader(valData, sampler=valSampler, batch_size=batch)
    
    model = BertForSequenceClassification.from_pretrained("bert-large-cased", num_labels=2)
    optimizerParameters = getParameters(model)
    optimizer = BertAdam(optimizerParameters, lr=2e-5, warmup=.1)
        
    model = trainModel(model, valDataLoader, optimizer)
    model = evalModel(model, valDataLoader, optimizer)
  

getBert(tweetsSchiz1, device)
